httpd: support parse HTTP/0.9 request.
As per RFC-1945, there are four kinds of HTTP message:

>   HTTP-message   = Simple-Request           ; HTTP/0.9 messages
>                  | Simple-Response
>                  | Full-Request             ; HTTP/1.0 messages
>                  | Full-Response

>   Simple-Request  = "GET" SP Request-URI CRLF
>   Simple-Response = [ Entity-Body ]

Here we add support to parse HTTP request of version 0.9. And notably,
we set request._version to "0.9" once a HTTP/0.9 request is successfully
parsed to fix the keep-alive bug introduced in commit dd9577d7e0 where
a HTTP/0.9 request will be keep-alived due to comparing request._version
(which is empty) with "0.9".

Signed-off-by: Jianyong Chen <baluschch@gmail.com>