websocket: fix server shutdown.(WIP)
The websocket can listen multiple addresses through listen(),
and the fiber runs in the background. But when server shuts down,
it only waits for the last future returned by listen(), which may
cause problems. Here we use gate to wait for all the background
fibers.

Besides, do_accepts and do_accept_one are renamed to accept and
accept_one, since methods with do_ prefix are usually internally
used by those without the prefix.

Signed-off-by: Jianyong Chen <baluschch@gmail.com>